update releaser boilerplate + encourage v2.0.0-rc1
A few changes:
- Now it will write the right boilerplate to `README.md` and
  `examples/rules_cc/WORKSPACE`.
- tightened the release tarball: it now will depend only on the file
  content, but not the specific SHA from which the release is created.
  This way we can cut a release and update the README of the release
  tarballs *in the same commit*.
- In an upcoming PR the release script will verify that boilerplate in
  `README.md` and `examples/rules_cc/WORKSPACE` are up to date. But we
  need to cut `v2.0.0-rc2` first; which I will do with this release
  script. I will also update the release process wiki with the added
  features of this tool.

The releaser tool is intended to be used in two modes:

1. *Making* the release: the target tag does not exist.
2. *CI*: the target tag does exist.

The release tool supports both. Second case is a bit more interesting:
it will create the release tarball, update the boilerplate from that
file, but will not run `git tag`. That way, the CI job can verify that
the README.md have unchanged. The CI is expected to run this with the
last git tag; that way we will be sure that the examples and the README
are up to date.