launcher: rename to zig-wrapper + passthrough mode
Also remove `ld64.lld`, since it was removed from zig upstream back in
2021: https://github.com/ziglang/zig/commit/0e15205521b9a8c95db3c1714dffe3be1df5cda1

Until today the "launcher" could do one thing well: set the environment
variables and exec the `zig c++` command. However, it was clumsy with
other subcommands: `ar`, `ld.lld` do not expect the `-target`. So there
was this "special case" for subcommands that are not c++, but, for the
sake of consistency, needed to be in the top-level repository.

This change leaves only one sub-command for the target, which truly
expects a `-target` flag: `c++`. From now:
- all other sub-commands required for the toolchain are in top-level.
- anything that is not needed for the toolhain can be directly used via
  zig-wrapper.

This simplifies the launcher quite a bit.