Replace `cp` with zig implementation: v2
Previously `cp` was invoked with zig:

    zig run cp.zig -- SRC DST

However, this was reverted in 73db33bab01a365a8be01e8586203fc2e95b9f09
due to races in `zig build-exe`, as described in
https://github.com/ziglang/zig/issues/15860

This approach brings back `cp.zig`, but this time the `cp` binary is
pre-compiled before invoking it.

That way Bazel will execute `zig build-exe` only once for that target,
that way not hitting the race. As a side-effect, the operation will be
faster, because `cp` will now be a single binary, as opposed to bringing
the whole zig_sdk together.