Testing bzlmod earlier (#119)
bzlmod requires a different flag to override external dependencies. We
need to parameterize that flag depending on whether an example has
bzlmod enabled. With this change, our CI covers the the case of a test
module depending on hermetic_cc_toolchain as a `bazel_dep`, pointing to
a release archive created from HEAD and running all tests in the test
module. Hopefully we can catch issues in our CI, rather than on BCR's CI
after a release is cut.

With this change, we no longer run the tests again in BCR's CI and no
longer need to include the example in the release archive. We only need
to verify that the latest release can be loaded by the [anonymous test
module](https://github.com/bazelbuild/bazel-central-registry/blob/main/docs/README.md#anonymous-module-test).