Add wasm32-wasi-musl target
I would like to be able to compile wasm/wasi with zig + bazel. This adds
a toolchain to be able to do so. Wasm/wasi is a pretty rapidly changing
space, but I've choosen the most specific naming that I could (WASI
preview 1).

I've confirmed this works by running the compiled binary with wazero,
a popular golang based wasm/wasi runtime in the tests.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>