Move toolchain registrations to .bazelrc
After the migration to bzlmod, the toolchain registration for the
module's code and tests ended up in the MODULE.bazel itself, which means
it gets carried across to module users. This means it's not possible to
choose which toolchains are wanted, and defeats some of the use cases
mentioned in README.md.

Moving this to .bazelrc is a workaround that restores the original
(pre-bzlmod) behavior.

Fixes #181