Append username to the default tmp directory
When two people build code on the same machine, there can be a collision. The second person may not be able to write to zig-cache if it was created by someone else. When it happens, the error message is quite confusing (see `_compile_failed` variable).

With this commit, each user should get a different directory by default. It's still possible to change the default using HERMETIC_CC_TOOLCHAIN_CACHE_PREFIX.

Signed-off-by: Laurent Le Brun <laurentlb@gmail.com>