feat: create cc_toolchains for multiple exec platforms (#196)
This change aims to allow hermetic_cc_toolchain to create toolchains for
multiple exec platforms. It now generates repositories of the format
{exec}-zig_sdk and registers each toolchain to only be
exec_compatible_with those specific platform constraints.

I want this PR to be a WIP as we discuss how we want this to look and if
you're happy with it's general flow.

This is an attempt at fixing
https://github.com/uber/hermetic_cc_toolchain/issues/148

So far I have tested that my local client (mac) can build c code using
the toolchains registered on a `x86_64` remote executor and have seen
success. It'll will be hard for me to test every combination; but I will
be able to local (x86_64) -> remote (arm64) tests as well.